requirements.txt
python-dotenv==0.19.0
google-cloud-bigquery==2.32.0